Fixed AssemblyBrowser::basesVisible() in case of right border. Fixes UGENE-558.
git-svn-id: https://local.ugene.unipro.ru/svn/ugene/trunk@1037 7ca3a678-163c-e011-b84f-001e8c0267ff

+++ src/plugins/assembly_browser/src/AssemblyDensityGraph.cpp
@@ -110,7 +110,7 @@ void AssemblyDensityGraph::sl_launchCoverageCalculation()
 {
     if(browser->areCellsVisible()) {
         qint64 start = browser->getXOffsetInAssembly();
-        int length= browser->basesVisible();
+        qint64 length = browser->basesVisible();
 
         CalcCoverageInfoTaskSettings settings;
         settings.model = model;
